Themes: Switch themes when user changes
This patch grabs the theme config for the new user and then
creates a ThemeChangeRequest from that config.  This request is
then processed and applied for the new user.

Secondary users will not be able to change themes at this time.
This patch helps pave the way for that ability and provides a better
user experience when switching users.

Change-Id: I6d7e2ab312b8e3e5c099d1e9e2e62892bead10da
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>