Marshmallow: If a wake key is disabled by the user, do not wake from doze.
Adapted for M by @KreAch3R

Currently, any wake key will wake the device from a doze, even if that key has not been enabled as a wake key in Settings.

If the device is 'dreaming' in the Doze state, check if the user has explicitly disabled the wake key (or never enabled the setting in the first place) before waking the device.

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>

Conflicts:
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>