Themes: Allow packages to be set as non-themeable
This patch allows for specific packages to be declared as non
themeable.  Packages that are non themeable will not have theme
resources attached to their assets when their Resources object
is created.

This patch also removes isThemeable from CompatibilityInfo as this
object did not correctly reflect that the package is not themeable.

Change-Id: Id34c63ec0c3ef7c69df083da63559e0720ce0018
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>